[DataEntry] Extract utilities
